+++ src/config/dictionary/dictionary.ts
@@ -10,7 +10,7 @@ import { MAGIC_ITEMS } from "./items/magicItems";
 import { SENSES } from "./actor/senses";
 import { SPELL } from "./spell/spell";
 import { PARSING_ACTIONS, PARSING_ATTACK_ACTIONS } from "./parsing/actions";
-import { FEATURE_SPELLS_IGNORE, IGNORE_SPELLS_GRANTED_BY_CLASS_FEATURES, IGNORE_SPELLS_GRANTED_BY_FEATS, LEVEL_SCALE, NO_GRANTED_SPELL_LIST_FEATURE_2014_INCLUDES, PARSING_CHOICE_FEATURES, PARSING_FEATURES } from "./parsing/features";
+import { FEATURE_SPELLS_IGNORE, FORCE_TRAIT_SPELL_ADVANCEMENT_ON_RACE, IGNORE_SPELLS_GRANTED_BY_CLASS_FEATURES, IGNORE_SPELLS_GRANTED_BY_FEATS, LEVEL_SCALE, NO_GRANTED_SPELL_LIST_FEATURE_2014_INCLUDES, PARSING_CHOICE_FEATURES, PARSING_FEATURES } from "./parsing/features";
 import { EXCLUDED_EFFECT_MODIFIERS } from "./effects/excluded";
 import { VISION_5E_EFFECTS } from "./effects/vision5e";
 import { RESETS } from "./actor/resets";
@@ -63,6 +63,7 @@ const DICTIONARY = {
       "2014": NO_GRANTED_SPELL_LIST_FEATURE_2014_INCLUDES,
     },
     levelScale: LEVEL_SCALE,
+    forceTraitSpellAdvancementOnRace: FORCE_TRAIT_SPELL_ADVANCEMENT_ON_RACE,
   },
   identifierAdjustments: IDENTIFIER_ADJUSTMENTS,
   effects: {